test-ep: Add an initial test endpoint
The rocm-axiio library uses the concepts of endpoints to refer to the
devices that talked to by the accelerator. As such we are implementing
a test endpoint that we can use for performance tuning, debug and
other things. This test endpoint is implemented in software and can
run on either an accelerator or on a CPU.

Commit a very initial version of the test endpoint plus some cmake
code to build the endpoint. Also commit new CI workflows for linting
and building of this code.